cloud-functions/verifier/analyser/requirements.txt
insightface==0.7.3
mediapipe==0.10.21
requests~=2.32.3
numpy~=1.26.4
opencv-python==4.11.0.86
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.7
functions-framework~=3.8.3
onnxruntime>=1.15.1
Flask~=3.1.1